REstore CHnages

+++ src/main/java/com/energytrade/app/dao/EventCustomerMappingDao.java
@@ -123,8 +123,13 @@ public class EventCustomerMappingDao extends AbstractBaseDao {
 				return response;
 			}
 			eventCustomerMappingRepo.withdrawFromEvent(userId, eventId);
+			if (evmt.getEventCustomerStatusId() == 3) {
 			eventRepository.removeEventPower(evmt.getCommitedPower(),eventId);
 			eventSetRepository.removeCommittedPower(evmt.getCommitedPower(), event.getAllEventSet().getEventSetId());
+			}
+			else if (evmt.getEventCustomerStatusId() == 4) {
+				eventCustomerMappingRepo.withdrawFromEventInCounterBid(userId, eventId,event.getExpectedPrice());
+			}
 			response.put("message", CustomMessages.getCustomMessages("SUC"));
 			response.put("key", "200");
 		} catch (Exception e) {
@@ -158,9 +163,11 @@ public class EventCustomerMappingDao extends AbstractBaseDao {
 		try {
 			EventCustomerMapping evtmp = eventCustomerMappingRepo.getEventCustomerMapping(eventId, userId);
 			double resultantPower = updatedCommitedPower - evtmp.getCommitedPower();
-			eventCustomerMappingRepo.updateEventCommitments(userId, eventId, resultantPower,
+			eventCustomerMappingRepo.updateEventCommitments(userId, eventId, updatedCommitedPower,
 					updatedCounterBidAmount);
-			eventRepository.addEventPower(updatedCommitedPower, eventId);
+			if (evtmp.getEventCustomerStatusId() == 3) {
+			eventRepository.addEventPower(resultantPower, eventId);
+			}
 			ArrayList<EventCustomerDevices> eventCustomerDevicesList = new ArrayList<EventCustomerDevices>();
 			EventCustomerMapping eventCustomerMapping = eventCustomerMappingRepo.getEventCustomerMapping(eventId,
 					userId);
